Properly obtained Dependencies of DtoCollection
First we need to get a list of fixed dependencies, that create
dependency graph from mandatory attributes. Which are attributes
present in unique manager_ref and attributes with presence
validation. Without these attributes we are not able to save
the record.

Second we need to know which attributes are causing which of the
dependencies in our DtoCollection.

And third we need a way to blacklist attributes which also
removes the dependencies.


(transferred from ManageIQ/manageiq@4944f0e653a6e8d9c04543d87717e61ed2c46b39)