Remove TypedIndex for DTO saving
TypedIndex was doing n+1 queries when key was a relation.
So typed index was remplaced by simple concatenation of all
keys of the unique index with ''. If some key is a relation, it's
transformed to a foreign key.

Also no need to hold a separate delete_index.


(transferred from ManageIQ/manageiq@3c1852a42327ae14ab9d57eb14e38eae0435cdcc)