Move edges when transforming nodes, due to the dependency cycle.
Move edges when transforming nodes, due to the dependency cycle.

If we remove an attribute that was a dependency of another node,
we need to move also the dependency. So e.g. floating_ip depends
on network_port's attribute vm, but we move that attribute to new
network_port dto_collection. We will need to move also the
dependency to point to the new dto_collection.


(transferred from ManageIQ/manageiq@3e31c540c54265e9dcd8a824473490510cbee805)