We need a custom_manager_uuid for DB based Dto with a relation in ref
The unique key list manager_ref contains a reference, which can't be
built automatically when loading the DtoCollection from the DB, we
need to provide a custom_manager_uuid lambda, that builds the correct
manager_uuid given a DB record.

An example for Disk:
The manager ref is [:hardware, :device_name]

With API based DtoCollections, the hardware gets automatically
converted to manager_ref of the :hardware DtoCollection.

But when loading from DB, the hardware is just a relation, so
we've lost the information about the DtoCollection under the hardware.

So we need to manually build the reference as:

[hardware.hardware.vm_or_template.ems_ref, hardware.device_name]


(transferred from ManageIQ/manageiq@743f19857a31edfc89c83cb271fdb9b3bee00cfa)