find_missing_in_local_db strategy
This strategy is comes handy when we do targeted refresh of
a inventory_collection while there is a lazy_link to record
present only in the DB.

E.g. we refresh a VM and Hardware, but hardware has a lazy link
to hardware owned by Vm's image.


(transferred from ManageIQ/manageiq@62125931304cf89561ed5fca74c53fec43781662)