Use strings for nested joins
In the following:

https://github.com/rails/rails/pull/25146

Rails added a regression to making nested joins with references to those
joins in the WHERE clause, and this causes an error when running this
test suite in Rails 5.0.1.

This has been addressed in Rails with:

https://github.com/rails/rails/pull/27598

And has been merged into master, but it is likely that Rails 5.0.1 will
not be pulled from Rubygems, and the ManageIQ will fail to run tests
properly with that version of Rails.

Because of that, and to allow the tests to function in the same way, the
simplest thing to do is to stringify the keys manually avoid the error
for now.  This is against the convention (most Rails devs prefer
symbols), but this allows the tests to remain as they were.


(transferred from ManageIQ/manageiq@4e96ee4fb928dc3e03e7c48dd37146005cc56427)