Add ApplicationRecordLite that is for holding is and class of AR
Add ApplicationRecordLite that is for holding is and class of AR.
ActiveRecord object are to huge and we need a query to obtain them,
in most cases we have a class and id, which is just enough
to fil la relation. No need to go to db.


(transferred from ManageIQ/manageiq@8e8a409f65c3fd22fe304f75fbd0f0e6489c1747)