Use proper multi select condition
Right now we are using this for loading from the db:
"col1 IN [a,b] AND col2 IN [e,f]"
which does give incorrect results since we need to load exact
combination:
"(col1, col2) IN [(a,e), (b,f), (b,e)]

The exact combination is especially needed when we are
determining if we should create/delete/update the records.

Unifying the comparation and loading from the db to the correct
multiselect query.


(transferred from ManageIQ/manageiq@78558b31f2b9acbf3a52e93b05b0daf73b85486e)