[rails6][spec/models/disk.rb] Fix .present
Due to a performance fix introduced in Rails 6:

  https://github.com/rails/rails/commit/cc2d614e

Columns named "present" end up with the following error:

  ActiveRecord::DangerousAttributeError:
    present? is defined by Active Record. Check to make sure that you
    don't have an attribute or method with the same name.

In manageiq, we kinda punted on fixing this column since it is hard to
know where it might be used elsewhere (Automate, etc.) so if a rename is
the final route we code with it is something that ideally happens in a
targeted effort outside the Rails 6 upgrade.  But here, since it is only
example models for specs, changing here isn't as dangerous so being
cavalier about the change isn't really an issue.