Fix the broken test matrix and drop old rubies and rails
The ci.yaml matrix of rubies and rails versions didn't match the case statement in the
Gemfile.  Rails 7.0 and 7.1 were not included in the case statement previously.

Now, we sync them up and make the following changes:

Drop:
* ruby 2.7
* rails 6.0
* rails 6.1

Add:
* ruby 3.3
* rails 7.2

We set a minimum ruby version of 3.0 and rails 7.0.8 and allow versions less than
8.0, so we can test with 7.1 and 7.2.